% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toon.R
\docType{package}
\name{toon}
\alias{toon-package}
\alias{toon}
\title{Token-Oriented Object Notation (TOON)}
\description{
An R implementation for converting R objects (vectors, lists, data frames)
into a custom, human-readable structured format called The Object Notation (TOON).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/laresbernardo/toon}
  \item Report bugs at \url{https://github.com/laresbernardo/toon/issues}
}

}
\author{
\strong{Maintainer}: Bernardo Lares \email{toon@bervos.org}

}
